moved existing qa samplers to the new qa sampler app
