+++ samples/sampler/stories/default/ImageItem.js
@@ -28,15 +28,3 @@ storiesOf('Agate', module)
 			text: 'Basic usage of ImageItem'
 		}
 	);
-
-storiesOf('Agate QA.Image', module)
-	.add(
-		'without children',
-		() => (
-			<div style={{width: ri.scaleToRem(400), height: ri.scaleToRem(300)}}>
-				<ImageItem
-					src="http://placehold.it/300x400/9037ab/ffffff&text=Image0"
-				/>
-			</div>
-		)
-	);
