+++ samples/sampler/stories/default/Item.js
@@ -40,34 +40,3 @@ storiesOf('Agate', module)
 			text: 'Basic usage of Item'
 		}
 	);
-
-storiesOf('Agate QA.Item', module)
-	.add(
-		'with icons and a text in slotAfter',
-		() => {
-			const slotBeforeIcon = select('slotBefore Icon', iconList, Config);
-			const slotAfter = [
-				select('slotAfter Icon 1', iconList, Config),
-				select('slotAfter Icon 2', iconList, Config),
-				text('slotAfter Text', Config, 'Sub Text')
-			].filter(subText => !!subText).map(subText => (iconList.indexOf(subText) > 0 ? <Icon key={subText}>{subText}</Icon> : subText));
-
-			return (
-				<Item
-					disabled={boolean('disabled', Config)}
-					inline={boolean('inline', Config)}
-					label={text('label', Config)}
-					labelPosition={select('labelPosition', ['above', 'after', 'before', 'below'], Config, 'below')}
-					selected={boolean('selected', Config)}
-				>
-					{slotBeforeIcon ? <slotBefore>
-						<Icon>{slotBeforeIcon}</Icon>
-					</slotBefore> : null}
-					{text('children', Config, 'Hello Item')}
-					{slotAfter.length > 0 ? <slotAfter>
-						{slotAfter}
-					</slotAfter> : null}
-				</Item>
-			);
-		}
-	);
