+++ samples/sampler/stories/qa/Item.js
@@ -1,17 +1,55 @@
+import {mergeComponentMetadata} from '@enact/storybook-utils';
+import {boolean, select, text} from '@enact/storybook-utils/addons/knobs';
+import UiItem, {ItemBase as UiItemBase} from '@enact/ui/Item';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 
 import Heading from '@enact/agate/Heading';
-import Icon from '@enact/agate/Icon';
-import Item from '@enact/agate/Item';
+import Icon, {icons} from '@enact/agate/Icon';
+import Item, {ItemBase} from '@enact/agate/Item';
+import Scroller from '@enact/agate/Scroller';
 
+const Config = mergeComponentMetadata('Item', UiItemBase, UiItem, ItemBase, Item);
 Item.displayName = 'Item';
 
-storiesOf('Agate QA.Item', module)
+const iconList = [null, ...Object.keys(icons)];
+
+storiesOf('Item', module)
+	.add(
+		'with icons and a text in slotAfter',
+		() => {
+			const slotBeforeIcon = select('slotBefore Icon', iconList, Config);
+			const slotAfter = [
+				select('slotAfter Icon 1', iconList, Config),
+				select('slotAfter Icon 2', iconList, Config),
+				text('slotAfter Text', Config, 'Sub Text')
+			].filter(subText => !!subText).map(subText => (iconList.indexOf(subText) > 0 ? <Icon key={subText}>{subText}</Icon> : subText));
+
+			return (
+				<Item
+					disabled={boolean('disabled', Config)}
+					inline={boolean('inline', Config)}
+					label={text('label', Config)}
+					labelPosition={select('labelPosition', ['above', 'after', 'before', 'below'], Config, 'below')}
+					selected={boolean('selected', Config)}
+				>
+					{slotBeforeIcon ? <slotBefore>
+						<Icon>{slotBeforeIcon}</Icon>
+					</slotBefore> : null}
+					{text('children', Config, 'Hello Item')}
+					{slotAfter.length > 0 ? <slotAfter>
+						{slotAfter}
+					</slotAfter> : null}
+				</Item>
+			);
+		}
+	);
+
+storiesOf('Item', module)
 	.add(
 		'Kitchen Sink',
 		() => (
-			<>
+			<Scroller>
 				<Heading showLine>Item variations</Heading>
 				<Item>Default Item</Item>
 				<Item label="label above" labelPosition="above">Item</Item>
@@ -54,7 +92,7 @@ storiesOf('Agate QA.Item', module)
 				<Item inline label="label below" labelPosition="below" slotAfter={<Icon>home</Icon>} slotBefore={<Icon>home</Icon>}>Item</Item>
 				<Item inline label="label before" labelPosition="before" slotAfter={<Icon>home</Icon>} slotBefore={<Icon>home</Icon>}>Item</Item>
 				<Item inline label="label after" labelPosition="after" slotAfter={<Icon>home</Icon>} slotBefore={<Icon>home</Icon>}>Item</Item>
-			</>
+			</Scroller>
 		),
 		{
 			text: 'Item Kitchen Sink'
